Django_REST/src/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-rest-auth==0.9.5
djangorestframework==3.11.0
future==0.18.2
idna==2.10
kombu==4.6.11
nba-api==1.1.8
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
